+++ mathf.h
@@ -3,16 +3,16 @@
     \author Gabriele Triglia
 */
 
-#ifndef MATHF_H 
+#ifndef MATHF_H
 #define MATHF_H
 
 #include <stdio.h>
-#include <stdlib.h> 
+#include <stdlib.h>
 #include <string.h>
 #include "structs.h"
 
-float Polynomial(poly_s polyf, float in); 
-void Rectangular(float* values, int size, float stepsize, float* integ1, float* integ2); 
-float Trapezoidal(float* values, int size, float stepsize); 
+double Polynomial(poly_s polyf, double in);
+void Rectangular(double *values, int size, double stepsize, double *integ1, double *integ2);
+double Trapezoidal(double *values, int size, double stepsize);
 
 #endif
